Megjavítottam a First és LastXItemsTransactiont, mert Dávid nem írta át csak a file nevét :|. A Databa írtam egy új függvényt ami kiszámolja a listák alapján az összpénzt. Lista feltöltést átraktam a MainActivitybe. Financebe van már grafikon aminek lehetetlen módosítani a háttérszínét (teljesen nyílvánvaló)

+++ app/src/main/java/com/androidhf/data/Transaction.kt
@@ -12,7 +12,7 @@ import java.time.LocalTime
 
 
 class Transaction(
-   _amount : Double,
+   _amount : Int,
    _reason : String,
    _date : LocalDate,
    _time: LocalTime,
